Remove updating pg_type for pgis_abs
This dummy datatype `pgis_abs` was removed as part of postgis 2.5, but
upstream postgis missed removing it from this script. This sql is
appended to the upgrade/unpackaged sql, so when trying to upgrade the
extension to 2.5, it would fail with the error `ERROR: permission
denied: "pg_type" is a system catalog`. Updating the sql to remove this
altogether as it is not required. This was also recently fixed in
upstream postgis.