Add ordered aggregate when upgrading
The upgrade script generates sql files that are executed when upgrading.
Since ORDERED AGGREGATES are supported for GPDB, there were a few
aggregates that were not included in the generated sql files. Updating
the script to match both `CREATE AGGREGATE`/`CREATE ORDERED AGGREGATE`
syntax and add those to the upgrade sql.