Update lookup PostGIS to replicated tables
In order for the PostGIS functions to work with GPDB, the lookup tables
need to be DISTRIBUTED REPLICATED as these tables are internally queried
when calling address/geocoder functions on table data.
This commit add test cases to ensure that executing these functions on a
table, returns results instead of erroring out.