Update the README.md and source files (#1)
We have updated the README file to provide a detailed guide to users on
how to build and install PostGIS for the Cloudberry Database. Currently,
we only support installation on CentOS, but we are planning to expand to
other systems in the future. Additionally, we have updated the PostGIS
source tar file.